@@ -5,8 +5,8 @@ import { JwtService } from '@nestjs/jwt';
 import { User } from '../entities/user.entity';
 import { OrganizationUsersService } from './organization_users.service';
 import { EmailService } from './email.service';
-const bcrypt = require('bcrypt');
-var uuid = require('uuid');
+import bcrypt from 'bcrypt';
+import uuid from 'uuid';
 
 @Injectable()
 export class AuthService {
@@ -15,13 +15,13 @@ export class AuthService {
     private jwtService: JwtService,
     private organizationsService: OrganizationsService,
     private organizationUsersService: OrganizationUsersService,
-    private emailService: EmailService,
-  ) { }
+    private emailService: EmailService
+  ) {}
 
   async validateUser(email: string, password: string): Promise<User> {
     const user = await this.usersService.findByEmail(email);
 
-    if(!user) return null;
+    if (!user) return null;
 
     const isVerified = await bcrypt.compare(password, user.password);
 
@@ -39,7 +39,7 @@ export class AuthService {
         email: user.email,
         first_name: user.firstName,
         last_name: user.lastName,
-        role: user.role
+        role: user.role,
       };
     } else {
       throw new UnauthorizedException('Invalid credentials');
@@ -48,13 +48,14 @@ export class AuthService {
 
   async signup(params: any) {
     // Check if the installation allows user signups
-    if(process.env.DISABLE_SIGNUPS === 'true') {
-      return {}
+    if (process.env.DISABLE_SIGNUPS === 'true') {
+      return {};
     }
-    
+
     const { email } = params;
     const organization = await this.organizationsService.create('Untitled organization');
     const user = await this.usersService.create({ email }, organization);
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     const organizationUser = await this.organizationUsersService.create(user, organization, 'admin');
 
     this.emailService.sendWelcomeEmail(user.email, user.firstName, user.invitationToken);
@@ -71,8 +72,8 @@ export class AuthService {
 
   async resetPassword(token: string, password: string) {
     const user = await this.usersService.findByPasswordResetToken(token);
-    if(!user) {
-      throw new NotFoundException('Invalid token')
+    if (!user) {
+      throw new NotFoundException('Invalid token');
     } else {
       this.usersService.update(user.id, { password, forgotPasswordToken: null });
     }
