@@ -1,10 +1,4 @@
-import {
-  Module,
-  OnApplicationBootstrap,
-  OnModuleInit,
-  RequestMethod,
-  MiddlewareConsumer,
-} from '@nestjs/common';
+import { Module, OnApplicationBootstrap, OnModuleInit, RequestMethod, MiddlewareConsumer } from '@nestjs/common';
 
 import { Connection } from 'typeorm';
 import { TypeOrmModule } from '@nestjs/typeorm';
@@ -25,7 +19,7 @@ import { AppController } from './controllers/app.controller';
 import { AppService } from './services/app.service';
 import { AuthModule } from './modules/auth/auth.module';
 import { UsersModule } from './modules/users/users.module';
-import { AppConfigModule } from './modules/app_config/app_config.module'
+import { AppConfigModule } from './modules/app_config/app_config.module';
 import { AppsModule } from './modules/apps/apps.module';
 import { FoldersModule } from './modules/folders/folders.module';
 import { FolderAppsModule } from './modules/folder_apps/folder_apps.module';
@@ -75,7 +69,7 @@ if (process.env.SERVE_CLIENT !== 'false') {
   imports.unshift(
     ServeStaticModule.forRoot({
       rootPath: join(__dirname, '../../../', 'frontend/build'),
-    }),
+    })
   );
 }
 
@@ -85,7 +79,7 @@ if (process.env.APM_VENDOR == 'sentry') {
       dsn: process.env.SENTRY_DNS,
       tracesSampleRate: 1.0,
       debug: !!process.env.SENTRY_DEBUG,
-    }),
+    })
   );
 }
 
