@@ -1,5 +1,5 @@
-import { Controller, Get, Request, Post, UseGuards } from '@nestjs/common';
-import { JwtAuthGuard } from '../modules/auth/jwt-auth.guard';
+import { Controller, Get, Request, Post } from '@nestjs/common';
+
 import { AuthService } from '../services/auth.service';
 
 @Controller()
@@ -19,19 +19,19 @@ export class AppController {
   @Post('/forgot_password')
   async forgotPassword(@Request() req) {
     await this.authService.forgotPassword(req.body.email);
-    return {}
+    return {};
   }
 
   @Post('/reset_password')
   async resetPassword(@Request() req) {
     const { token, password } = req.body;
     await this.authService.resetPassword(token, password);
-    return {}
+    return {};
   }
 
   @Get('/health')
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   async healthCheck(@Request() req) {
     return { works: 'yeah' };
   }
-
 }
