@@ -1,4 +1,5 @@
-import { Controller, Get, Request } from '@nestjs/common';
+/* eslint-disable @typescript-eslint/no-unused-vars */
+import { Controller, Get } from '@nestjs/common';
 import { AppConfigService } from '@services/app_config.service';
 
 @Controller('config')
