@@ -1,4 +1,4 @@
-import { Controller, ForbiddenException, Get, Param, Post, Put, Query, Request, UnauthorizedException, UseGuards } from '@nestjs/common';
+import { Controller, ForbiddenException, Post, Request, UseGuards } from '@nestjs/common';
 import { JwtAuthGuard } from '../../src/modules/auth/jwt-auth.guard';
 import { decamelizeKeys } from 'humps';
 import { AppsAbilityFactory } from 'src/modules/casl/abilities/apps-ability.factory';
@@ -7,12 +7,11 @@ import { AppsService } from '@services/apps.service';
 
 @Controller('app_users')
 export class AppUsersController {
-
   constructor(
     private appUsersService: AppUsersService,
     private appsService: AppsService,
     private appsAbilityFactory: AppsAbilityFactory
-  ) { }
+  ) {}
 
   @UseGuards(JwtAuthGuard)
   @Post()
@@ -25,12 +24,11 @@ export class AppUsersController {
     const app = await this.appsService.find(appId);
     const ability = await this.appsAbilityFactory.appsActions(req.user, {});
 
-    if(!ability.can('createUsers', app)) {
+    if (!ability.can('createUsers', app)) {
       throw new ForbiddenException('you do not have permissions to perform this action');
     }
-    
+
     const appUser = await this.appUsersService.create(req.user, appId, organizationUserId, role);
     return decamelizeKeys(appUser);
   }
-
 }
