@@ -1,4 +1,4 @@
-import { Bind, Controller, ForbiddenException, Get, Param, Post, Put, Query, Req, Request, UseGuards } from '@nestjs/common';
+import { Controller, ForbiddenException, Get, Param, Post, Put, Query, Request, UseGuards } from '@nestjs/common';
 import { JwtAuthGuard } from '../../src/modules/auth/jwt-auth.guard';
 import { decamelizeKeys } from 'humps';
 import { DataSourcesService } from '../../src/services/data_sources.service';
@@ -8,27 +8,25 @@ import { DataQueriesService } from '@services/data_queries.service';
 
 @Controller('data_sources')
 export class DataSourcesController {
-
   constructor(
     private appsService: AppsService,
     private appsAbilityFactory: AppsAbilityFactory,
     private dataSourcesService: DataSourcesService,
     private dataQueriesService: DataQueriesService
-  ) { }
+  ) {}
 
   @UseGuards(JwtAuthGuard)
   @Get()
   async index(@Request() req, @Query() query) {
-
     const app = await this.appsService.find(query.app_id);
     const ability = await this.appsAbilityFactory.appsActions(req.user, {});
 
-    if(!ability.can('getDataSources', app)) {
+    if (!ability.can('getDataSources', app)) {
       throw new ForbiddenException('you do not have permissions to perform this action');
     }
-    
+
     const dataSources = await this.dataSourcesService.all(req.user, query.app_id);
-    let response = decamelizeKeys({ data_sources: dataSources });
+    const response = decamelizeKeys({ data_sources: dataSources });
 
     return response;
   }
@@ -42,10 +40,10 @@ export class DataSourcesController {
     const app = await this.appsService.find(appId);
     const ability = await this.appsAbilityFactory.appsActions(req.user, {});
 
-    if(!ability.can('createDataSource', app)) {
+    if (!ability.can('createDataSource', app)) {
       throw new ForbiddenException('you do not have permissions to perform this action');
     }
-    
+
     const dataSource = await this.dataSourcesService.create(req.user, name, kind, options, appId);
     return decamelizeKeys(dataSource);
   }
@@ -61,10 +59,10 @@ export class DataSourcesController {
     const app = await this.appsService.find(dataSource.appId);
     const ability = await this.appsAbilityFactory.appsActions(req.user, {});
 
-    if(!ability.can('updateDataSource', app)) {
+    if (!ability.can('updateDataSource', app)) {
       throw new ForbiddenException('you do not have permissions to perform this action');
     }
-    
+
     const result = await this.dataSourcesService.update(req.user, dataSourceId, name, options);
     return decamelizeKeys(result);
   }
@@ -94,11 +92,10 @@ export class DataSourcesController {
     const app = await this.appsService.find(dataSource.appId);
     const ability = await this.appsAbilityFactory.appsActions(req.user, {});
 
-    if(!ability.can('authorizeOauthForSource', app)) {
+    if (!ability.can('authorizeOauthForSource', app)) {
       throw new ForbiddenException('you do not have permissions to perform this action');
     }
 
     return await this.dataQueriesService.authorizeOauth2(dataSource, code);
   }
-
 }
