@@ -1,29 +1,36 @@
-import { Entity, Column, PrimaryGeneratedColumn, CreateDateColumn, UpdateDateColumn, ManyToOne, JoinColumn, BaseEntity } from 'typeorm';
+import {
+  Entity,
+  Column,
+  PrimaryGeneratedColumn,
+  CreateDateColumn,
+  UpdateDateColumn,
+  ManyToOne,
+  JoinColumn,
+  BaseEntity,
+} from 'typeorm';
 import { App } from './app.entity';
 
-@Entity({ name: "app_versions" })
+@Entity({ name: 'app_versions' })
 export class AppVersion extends BaseEntity {
-
-  @PrimaryGeneratedColumn("uuid")
+  @PrimaryGeneratedColumn('uuid')
   id: string;
 
-  @Column( { name: 'name' } )
+  @Column({ name: 'name' })
   name: string;
 
-  @Column("simple-json", { name: 'definition' } )
+  @Column('simple-json', { name: 'definition' })
   definition;
 
-  @Column( { name: 'app_id' } )
+  @Column({ name: 'app_id' })
   appId: string;
 
   @CreateDateColumn({ default: () => 'now()', name: 'created_at' })
   createdAt: Date;
-  
+
   @UpdateDateColumn({ default: () => 'now()', name: 'updated_at' })
   updatedAt: Date;
 
-  @ManyToOne(() => App, appVersion => appVersion.id)
-  @JoinColumn({ name: "app_id" })
+  @ManyToOne(() => App, (appVersion) => appVersion.id)
+  @JoinColumn({ name: 'app_id' })
   app: App;
-  
 }
