@@ -1,11 +1,19 @@
-import { Entity, Column, PrimaryGeneratedColumn, CreateDateColumn, UpdateDateColumn, ManyToOne, JoinColumn, BaseEntity } from 'typeorm';
+import {
+  Entity,
+  Column,
+  PrimaryGeneratedColumn,
+  CreateDateColumn,
+  UpdateDateColumn,
+  ManyToOne,
+  JoinColumn,
+  BaseEntity,
+} from 'typeorm';
 import { Organization } from './organization.entity';
 import { User } from './user.entity';
 
-@Entity({ name: "organization_users" })
+@Entity({ name: 'organization_users' })
 export class OrganizationUser extends BaseEntity {
-
-  @PrimaryGeneratedColumn("uuid")
+  @PrimaryGeneratedColumn('uuid')
   id: string;
 
   @Column()
@@ -14,24 +22,23 @@ export class OrganizationUser extends BaseEntity {
   @Column({ type: 'enum', enumName: 'status', enum: ['invited', 'active', 'archived'] })
   status: string;
 
-  @Column({ name: 'organization_id' }) 
-  organizationId: string
+  @Column({ name: 'organization_id' })
+  organizationId: string;
 
-  @Column({ name: 'user_id' }) 
-  userId: string
+  @Column({ name: 'user_id' })
+  userId: string;
 
   @CreateDateColumn({ default: () => 'now()', name: 'created_at' })
   createdAt: Date;
-  
+
   @UpdateDateColumn({ default: () => 'now()', name: 'updated_at' })
   updatedAt: Date;
 
-  @ManyToOne(() => Organization, organization => organization.id)
-  @JoinColumn({ name: "organization_id" })
-  organization: Organization;  
+  @ManyToOne(() => Organization, (organization) => organization.id)
+  @JoinColumn({ name: 'organization_id' })
+  organization: Organization;
 
-  @ManyToOne(() => User, user => user.id)
-  @JoinColumn({ name: "user_id" })
-  user: User;  
-
+  @ManyToOne(() => User, (user) => user.id)
+  @JoinColumn({ name: 'user_id' })
+  user: User;
 }
