@@ -1,6 +1,6 @@
-import { QueryError } from "src/modules/data_sources/query.error";
+import { QueryError } from 'src/modules/data_sources/query.error';
 
-export function parseJson(jsonString: string, errorMessage?: string): object {
+export function parseJson(jsonString: string, errorMessage?: string): Record<string, any> {
   try {
     return JSON.parse(jsonString);
   } catch (err) {
@@ -14,16 +14,15 @@ export async function cacheConnection(dataSourceId: string, connection: any): Pr
 }
 
 export async function getCachedConnection(dataSourceId, dataSourceUpdatedAt): Promise<any> {
-
   const cachedData = globalThis.CACHED_CONNECTIONS[dataSourceId] || {};
 
-  if(cachedData) {
+  if (cachedData) {
     const updatedAt = new Date(dataSourceUpdatedAt || null);
     const cachedAt = new Date(cachedData.updatedAt || null);
 
-    const diffTime = (cachedAt.getTime() - updatedAt.getTime())/1000;
+    const diffTime = (cachedAt.getTime() - updatedAt.getTime()) / 1000;
 
-    if(diffTime < 0) { 
+    if (diffTime < 0) {
       return null;
     } else {
       return cachedData['connection'];
