@@ -1,5 +1,4 @@
 import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
 import { AppConfigController } from '@controllers/app_config.controller';
 import { AppConfigService } from '@services/app_config.service';
 
@@ -8,5 +7,4 @@ import { AppConfigService } from '@services/app_config.service';
   imports: [],
   providers: [AppConfigService],
 })
-
-export class AppConfigModule{}
+export class AppConfigModule {}
