@@ -2,5 +2,4 @@ import { SetMetadata } from '@nestjs/common';
 import { PolicyHandler } from './policyhandler.interface';
 
 export const CHECK_POLICIES_KEY = 'check_policy';
-export const CheckPolicies = (...handlers: PolicyHandler[]) =>
-  SetMetadata(CHECK_POLICIES_KEY, handlers);
+export const CheckPolicies = (...handlers: PolicyHandler[]) => SetMetadata(CHECK_POLICIES_KEY, handlers);
