@@ -1,5 +1,5 @@
 export type ConnectionTestResult = {
-    status: 'ok' | 'failed',
-    message?: string,
-    data?: object
-}
+  status: 'ok' | 'failed';
+  message?: string;
+  data?: Record<string, any>;
+};
