@@ -1,5 +1,5 @@
 export type QueryResult = {
-    status: 'ok' | 'failed' | 'needs_oauth'
-    errorMessage?: string,
-    data: Array<object> | object
-}
+  status: 'ok' | 'failed' | 'needs_oauth';
+  errorMessage?: string;
+  data: Array<Record<string, any>> | Record<string, any>;
+};
