@@ -10,29 +10,18 @@ export class AppConfigService {
       : this.fetchDefaultConfig();
 
     const mapEntries = await Promise.all(
-      whitelistedConfigVars.map(
-        (envVar) => [envVar, process.env[envVar]] as [string, string],
-      ),
+      whitelistedConfigVars.map((envVar) => [envVar, process.env[envVar]] as [string, string])
     );
 
     return Object.fromEntries(mapEntries);
   }
 
   fetchDefaultConfig() {
-    return [
-      'TOOLJET_SERVER_URL',
-      'RELEASE_VERSION',
-      'GOOGLE_MAPS_API_KEY',
-      'APM_VENDOR',
-      'SENTRY_DNS',
-      'SENTRY_DEBUG',
-    ];
+    return ['TOOLJET_SERVER_URL', 'RELEASE_VERSION', 'GOOGLE_MAPS_API_KEY', 'APM_VENDOR', 'SENTRY_DNS', 'SENTRY_DEBUG'];
   }
 
   fetchAllowedConfigFromEnv() {
-    const whitelistedConfigVars = process.env.ALLOWED_CLIENT_CONFIG_VARS.split(
-      ',',
-    ).map((envVar) => envVar.trim());
+    const whitelistedConfigVars = process.env.ALLOWED_CLIENT_CONFIG_VARS.split(',').map((envVar) => envVar.trim());
 
     return whitelistedConfigVars;
   }
