@@ -3,16 +3,15 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Credential } from '../../src/entities/credential.entity';
 import { Repository } from 'typeorm';
 import { EncryptionService } from './encryption.service';
-const crypto = require('crypto')
+// import crypto from 'crypto';
 
 @Injectable()
 export class CredentialsService {
-
   constructor(
     private encryptionService: EncryptionService,
     @InjectRepository(Credential)
-    private credentialsRepository: Repository<Credential>,
-  ) { }
+    private credentialsRepository: Repository<Credential>
+  ) {}
 
   async create(value: string): Promise<Credential> {
     const newCredential = this.credentialsRepository.create({
@@ -24,9 +23,13 @@ export class CredentialsService {
     return credential;
   }
 
-  async getValue(credentialId: string): Promise<string> { 
+  async getValue(credentialId: string): Promise<string> {
     const credential = await this.credentialsRepository.findOne(credentialId);
-    const decryptedValue =  await this.encryptionService.decryptColumnValue('credentials', 'value', credential.valueCiphertext)
+    const decryptedValue = await this.encryptionService.decryptColumnValue(
+      'credentials',
+      'value',
+      credential.valueCiphertext
+    );
     return decryptedValue;
   }
 }
