@@ -8,49 +8,48 @@ import { Folder } from '../entities/folder.entity';
 
 @Injectable()
 export class FoldersService {
-
   constructor(
     @InjectRepository(Folder)
     private foldersRepository: Repository<Folder>,
     @InjectRepository(FolderApp)
     private folderAppsRepository: Repository<FolderApp>,
     @InjectRepository(App)
-    private appsRepository: Repository<App>,
-  ) { }
+    private appsRepository: Repository<App>
+  ) {}
 
   async create(user: User, folderName): Promise<Folder> {
-    return this.foldersRepository.save(this.foldersRepository.create({
+    return this.foldersRepository.save(
+      this.foldersRepository.create({
         name: folderName,
         createdAt: new Date(),
         updatedAt: new Date(),
         organizationId: user.organizationId,
-    }));
+      })
+    );
   }
 
   async all(user: User): Promise<Folder[]> {
-
     return await this.foldersRepository.find({
-        where: {
-            organizationId: user.organizationId,
-        },
-        relations: ['folderApps'],
-        order: {
-            name: 'ASC'
-        }
+      where: {
+        organizationId: user.organizationId,
+      },
+      relations: ['folderApps'],
+      order: {
+        name: 'ASC',
+      },
     });
   }
 
-  async findOne(folderId: string ): Promise<Folder> {
+  async findOne(folderId: string): Promise<Folder> {
     return await this.foldersRepository.findOneOrFail(folderId);
   }
 
   async userAppCount(user: User, folder: Folder) {
     const result = await this.foldersRepository
       .createQueryBuilder('folder')
-      .where("id = :id", { id: folder.id })
-      .loadRelationCountAndMap(
-        'folder.appCount', 'folder.apps', 'apps',
-        qb => qb.andWhere("apps.user_id = :user_id", { user_id: user.id })
+      .where('id = :id', { id: folder.id })
+      .loadRelationCountAndMap('folder.appCount', 'folder.apps', 'apps', (qb) =>
+        qb.andWhere('apps.user_id = :user_id', { user_id: user.id })
       )
       .getMany();
 
@@ -60,21 +59,24 @@ export class FoldersService {
   async getAppsFor(user: User, folder: Folder, page: number): Promise<App[]> {
     const folderApps = await this.folderAppsRepository.find({
       where: {
-        folderId: folder.id
-      }
+        folderId: folder.id,
+      },
     });
 
-    const apps = await this.appsRepository.findByIds(folderApps.map(folderApp => folderApp.appId), {
-      where: {
-        user
-      },
-      relations: ['user'],
-      take: 10,
-      skip: 10 * (page - 1),
-      order: {
-        createdAt: 'DESC'
+    const apps = await this.appsRepository.findByIds(
+      folderApps.map((folderApp) => folderApp.appId),
+      {
+        where: {
+          user,
+        },
+        relations: ['user'],
+        take: 10,
+        skip: 10 * (page - 1),
+        order: {
+          createdAt: 'DESC',
+        },
       }
-    });
+    );
 
     return apps;
   }
