@@ -10,64 +10,69 @@ import { EmailService } from './email.service';
 
 @Injectable()
 export class OrganizationUsersService {
-
   constructor(
     @InjectRepository(OrganizationUser)
     private organizationUsersRepository: Repository<OrganizationUser>,
     private usersService: UsersService,
-    private emailService: EmailService,
-  ) { }
+    private emailService: EmailService
+  ) {}
 
   async findOne(id: string): Promise<OrganizationUser> {
     return await this.organizationUsersRepository.findOne({ id: id });
   }
 
   async inviteNewUser(currentUser: User, params: any): Promise<OrganizationUser> {
-
-    const userParams = <User> {
+    const userParams = <User>{
       firstName: params['first_name'],
       lastName: params['last_name'],
-      email: params['email']
-    }
+      email: params['email'],
+    };
 
     const user = await this.usersService.create(userParams, currentUser.organization);
     const organizationUser = await this.create(user, currentUser.organization, params.role);
 
-    this.emailService.sendOrganizationUserWelcomeEmail(user.email, user.firstName, currentUser.firstName, user.invitationToken);
+    this.emailService.sendOrganizationUserWelcomeEmail(
+      user.email,
+      user.firstName,
+      currentUser.firstName,
+      user.invitationToken
+    );
 
     return organizationUser;
   }
 
   async create(user: User, organization: Organization, role: string): Promise<OrganizationUser> {
-    return await this.organizationUsersRepository.save(this.organizationUsersRepository.create({
-      user,
-      organization,
-      role,
-      createdAt: new Date(),
-      updatedAt: new Date(),
-  }));
+    return await this.organizationUsersRepository.save(
+      this.organizationUsersRepository.create({
+        user,
+        organization,
+        role,
+        createdAt: new Date(),
+        updatedAt: new Date(),
+      })
+    );
   }
 
   async changeRole(user: User, id: string, role: string) {
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     const organizationUser = await this.organizationUsersRepository.findOne(id);
     return await this.organizationUsersRepository.update(id, { role });
   }
 
   async archive(id: string) {
-
     const organizationUser = await this.organizationUsersRepository.findOne(id);
 
-    if(organizationUser.role === 'admin') {
+    if (organizationUser.role === 'admin') {
       // Check if this is the last admin of the org
       const adminsCount = await this.organizationUsersRepository.count({
         where: {
           organizationId: organizationUser.organizationId,
           role: 'admin',
-          status: 'active'
-        }
+          status: 'active',
+        },
       });
 
-      if(adminsCount === 1) {
+      if (adminsCount === 1) {
         throw new BadRequestException('You cannot archive this user as there are no other active admin users.');
       }
     }
