@@ -1,6 +1,13 @@
 import * as request from 'supertest';
 import { INestApplication } from '@nestjs/common';
-import { authHeaderForUser, clearDB, createApplication, createUser, createNestAppInstance, createApplicationVersion, createDataQuery, createDataSource } from '../test.helper';
+import {
+  authHeaderForUser,
+  clearDB,
+  createApplication,
+  createUser,
+  createNestAppInstance,
+  createDataSource,
+} from '../test.helper';
 
 describe('data sources controller', () => {
   let app: INestApplication;
@@ -14,58 +21,77 @@ describe('data sources controller', () => {
   });
 
   it('should be able to create data sources of an app only if admin/developer of same organization', async () => {
-
     const adminUserData = await createUser(app, { email: 'admin@tooljet.io', role: 'admin' });
-    const developerUserData = await createUser(app, { email: 'developer@tooljet.io', role: 'developer', organization: adminUserData.organization });
-    const viewerUserData = await createUser(app, { email: 'viewer@tooljet.io', role: 'viewer', organization: adminUserData.organization });
+    const developerUserData = await createUser(app, {
+      email: 'developer@tooljet.io',
+      role: 'developer',
+      organization: adminUserData.organization,
+    });
+    const viewerUserData = await createUser(app, {
+      email: 'viewer@tooljet.io',
+      role: 'viewer',
+      organization: adminUserData.organization,
+    });
     const anotherOrgAdminUserData = await createUser(app, { email: 'another@tooljet.io', role: 'admin' });
     const application = await createApplication(app, { name: 'name', user: adminUserData.user });
 
-    const dataSourceParams = { 
-      name: 'name', 
-      options: [], 
-      kind: 'postgres', 
-      app_id: application.id
-    }
+    const dataSourceParams = {
+      name: 'name',
+      options: [],
+      kind: 'postgres',
+      app_id: application.id,
+    };
 
-    for(const userData of [adminUserData, developerUserData]) {
+    for (const userData of [adminUserData, developerUserData]) {
       const response = await request(app.getHttpServer())
         .post(`/data_sources`)
         .set('Authorization', authHeaderForUser(userData.user))
-        .send(dataSourceParams)
+        .send(dataSourceParams);
 
       expect(response.statusCode).toBe(201);
     }
 
     // Should not update if viewer or if user of another org
-    for(const userData of [anotherOrgAdminUserData, viewerUserData]) {
+    for (const userData of [anotherOrgAdminUserData, viewerUserData]) {
       const response = await request(app.getHttpServer())
         .post(`/data_sources`)
         .set('Authorization', authHeaderForUser(userData.user))
-        .send(dataSourceParams)
+        .send(dataSourceParams);
 
       expect(response.statusCode).toBe(403);
     }
-
   });
 
   it('should be able to update data sources of an app only if admin/developer of same organization', async () => {
-
     const adminUserData = await createUser(app, { email: 'admin@tooljet.io', role: 'admin' });
-    const developerUserData = await createUser(app, { email: 'developer@tooljet.io', role: 'developer', organization: adminUserData.organization });
-    const viewerUserData = await createUser(app, { email: 'viewer@tooljet.io', role: 'viewer', organization: adminUserData.organization });
+    const developerUserData = await createUser(app, {
+      email: 'developer@tooljet.io',
+      role: 'developer',
+      organization: adminUserData.organization,
+    });
+    const viewerUserData = await createUser(app, {
+      email: 'viewer@tooljet.io',
+      role: 'viewer',
+      organization: adminUserData.organization,
+    });
     const anotherOrgAdminUserData = await createUser(app, { email: 'another@tooljet.io', role: 'admin' });
     const application = await createApplication(app, { name: 'name', user: adminUserData.user });
-    const dataSource = await createDataSource(app, { name: 'name', options: [], kind: 'postgres', application: application, user: adminUserData.user });
+    const dataSource = await createDataSource(app, {
+      name: 'name',
+      options: [],
+      kind: 'postgres',
+      application: application,
+      user: adminUserData.user,
+    });
 
-    for(const userData of [adminUserData, developerUserData]) {
-      const newOptions = [ { key: 'email', value: userData.user.email } ]
+    for (const userData of [adminUserData, developerUserData]) {
+      const newOptions = [{ key: 'email', value: userData.user.email }];
       const response = await request(app.getHttpServer())
         .put(`/data_sources/${dataSource.id}`)
         .set('Authorization', authHeaderForUser(userData.user))
         .send({
-          options: newOptions
-        })
+          options: newOptions,
+        });
 
       expect(response.statusCode).toBe(200);
       await dataSource.reload();
@@ -73,13 +99,13 @@ describe('data sources controller', () => {
     }
 
     // Should not update if viewer or if user of another org
-    for(const userData of [anotherOrgAdminUserData, viewerUserData]) {
+    for (const userData of [anotherOrgAdminUserData, viewerUserData]) {
       const response = await request(app.getHttpServer())
         .put(`/data_sources/${dataSource.id}`)
         .set('Authorization', authHeaderForUser(userData.user))
         .send({
-          options: [ ]
-        })
+          options: [],
+        });
 
       expect(response.statusCode).toBe(403);
     }
@@ -87,16 +113,24 @@ describe('data sources controller', () => {
 
   it('should be able to list (get) datasources for an app only if admin/developer of same organization', async () => {
     const adminUserData = await createUser(app, { email: 'admin@tooljet.io', role: 'admin' });
-    const developerUserData = await createUser(app, { email: 'developer@tooljet.io', role: 'developer', organization: adminUserData.organization });
-    const viewerUserData = await createUser(app, { email: 'viewer@tooljet.io', role: 'viewer', organization: adminUserData.organization });
+    const developerUserData = await createUser(app, {
+      email: 'developer@tooljet.io',
+      role: 'developer',
+      organization: adminUserData.organization,
+    });
+    const viewerUserData = await createUser(app, {
+      email: 'viewer@tooljet.io',
+      role: 'viewer',
+      organization: adminUserData.organization,
+    });
     const application = await createApplication(app, { name: 'name', user: adminUserData.user });
     const anotherOrgAdminUserData = await createUser(app, { email: 'another@tooljet.io', role: 'admin' });
-    const dataSource = await createDataSource(app, { name: 'name', kind: 'postgres', application: application, user: adminUserData.user });
+    // const dataSource = await createDataSource(app, { name: 'name', kind: 'postgres', application: application, user: adminUserData.user });
 
-    for(const userData of [adminUserData, developerUserData, viewerUserData]) {
+    for (const userData of [adminUserData, developerUserData, viewerUserData]) {
       const response = await request(app.getHttpServer())
         .get(`/data_sources?app_id=${application.id}`)
-        .set('Authorization', authHeaderForUser(userData.user))
+        .set('Authorization', authHeaderForUser(userData.user));
 
       expect(response.statusCode).toBe(200);
       expect(response.body.data_sources.length).toBe(1);
@@ -104,28 +138,32 @@ describe('data sources controller', () => {
 
     // Forbidden if user of another organization
     const response = await request(app.getHttpServer())
-        .get(`/data_sources?app_id=${application.id}`)
-        .set('Authorization', authHeaderForUser(anotherOrgAdminUserData.user))
+      .get(`/data_sources?app_id=${application.id}`)
+      .set('Authorization', authHeaderForUser(anotherOrgAdminUserData.user));
 
-      expect(response.statusCode).toBe(403);
+    expect(response.statusCode).toBe(403);
   });
 
   it('should not be able to authorize OAuth code for a REST API source if user of another organization', async () => {
-
     const adminUserData = await createUser(app, { email: 'admin@tooljet.io', role: 'admin' });
     const anotherOrgAdminUserData = await createUser(app, { email: 'another@tooljet.io', role: 'admin' });
     const application = await createApplication(app, { name: 'name', user: adminUserData.user });
-    const dataSource = await createDataSource(app, { name: 'name', options: [], kind: 'restapi', application: application, user: adminUserData.user });
+    const dataSource = await createDataSource(app, {
+      name: 'name',
+      options: [],
+      kind: 'restapi',
+      application: application,
+      user: adminUserData.user,
+    });
 
     // Should not update if user of another org
     const response = await request(app.getHttpServer())
       .post(`/data_sources/${dataSource.id}/authorize_oauth2`)
       .set('Authorization', authHeaderForUser(anotherOrgAdminUserData.user))
       .send({
-        code: 'oauth-auth-code'
-      })
+        code: 'oauth-auth-code',
+      });
 
     expect(response.statusCode).toBe(403);
   });
-
 });
