@@ -18,9 +18,8 @@ describe('organizations controller', () => {
   });
 
   it('should list organization users', async () => {
-
     const userData = await createUser(app, { email: 'admin@tooljet.io', role: 'admin' });
-    const { organization, user, orgUser } = userData;
+    const { user, orgUser } = userData;
 
     const response = await request(app.getHttpServer())
       .get('/organizations/users')
@@ -32,17 +31,16 @@ describe('organizations controller', () => {
     await orgUser.reload();
 
     expect(response.body.users[0]).toStrictEqual({
-      email: user.email, 
-      first_name: user.firstName, 
-      id: orgUser.id, 
-      last_name: user.lastName, 
-      name: `${user.firstName} ${user.lastName}`, 
-      role: orgUser.role, 
-      status: orgUser.status
-    })
+      email: user.email,
+      first_name: user.firstName,
+      id: orgUser.id,
+      last_name: user.lastName,
+      name: `${user.firstName} ${user.lastName}`,
+      role: orgUser.role,
+      status: orgUser.status,
+    });
   });
 
-  
   afterAll(async () => {
     await app.close();
   });
