@@ -14,14 +14,11 @@ import { DataQuery } from 'src/entities/data_query.entity';
 import { DataSource } from 'src/entities/data_source.entity';
 
 export async function createNestAppInstance() {
-  let app: INestApplication;
-
   const moduleRef = await Test.createTestingModule({
     imports: [AppModule],
-  })
-  .compile();
+  }).compile();
 
-  app = moduleRef.createNestApplication();
+  const app: INestApplication = moduleRef.createNestApplication();
   await app.init();
 
   return app;
@@ -29,7 +26,7 @@ export async function createNestAppInstance() {
 
 export function authHeaderForUser(user: any) {
   const configService = new ConfigService();
-  const jwtService = new JwtService({secret: configService.get<string>('SECRET_KEY_BASE')});
+  const jwtService = new JwtService({ secret: configService.get<string>('SECRET_KEY_BASE') });
   const authPayload = { username: user.id, sub: user.email };
   const authToken = jwtService.sign(authPayload);
   return `Bearer ${authToken}`;
@@ -44,105 +41,121 @@ export async function clearDB() {
 }
 
 export async function createApplication(app, { name, user, isPublic }: any) {
-  let appRepository: Repository<App>;
-  appRepository = app.get('AppRepository');
-  let appUsersRepository: Repository<AppUser>;
-  appUsersRepository = app.get('AppUserRepository');
+  const appRepository: Repository<App> = app.get('AppRepository');
+  // appRepository = app.get('AppRepository');
+  const appUsersRepository: Repository<AppUser> = app.get('AppUserRepository');
+  // appUsersRepository = app.get('AppUserRepository');
 
-  user = user || await (await createUser(app, {})).user;
+  user = user || (await (await createUser(app, {})).user);
 
-  const newApp = await appRepository.save(appRepository.create({
-    name,
-    user,
-    isPublic: isPublic || false,
-    organizationId: user.organization.id,
-    createdAt: new Date(),
-    updatedAt: new Date()
-  }));
+  const newApp = await appRepository.save(
+    appRepository.create({
+      name,
+      user,
+      isPublic: isPublic || false,
+      organizationId: user.organization.id,
+      createdAt: new Date(),
+      updatedAt: new Date(),
+    })
+  );
 
-  await appUsersRepository.save(appUsersRepository.create({
-    app: newApp, 
-    user,
-    role: 'admin',
-    createdAt: new Date(),
-    updatedAt: new Date()
-  }));
+  await appUsersRepository.save(
+    appUsersRepository.create({
+      app: newApp,
+      user,
+      role: 'admin',
+      createdAt: new Date(),
+      updatedAt: new Date(),
+    })
+  );
 
   return newApp;
 }
 
 export async function createApplicationVersion(app, application) {
-  let appVersionsRepository: Repository<AppVersion>;
-  appVersionsRepository = app.get('AppVersionRepository');
+  const appVersionsRepository: Repository<AppVersion> = app.get('AppVersionRepository');
+  // appVersionsRepository = app.get('AppVersionRepository');
 
-  return await appVersionsRepository.save(appVersionsRepository.create({
-    app: application,
-    name: 'v0'
-  }));
+  return await appVersionsRepository.save(
+    appVersionsRepository.create({
+      app: application,
+      name: 'v0',
+    })
+  );
 }
 
 export async function createUser(app, { firstName, lastName, email, role, organization }: any) {
-  let userRepository: Repository<User>;
-  let organizationRepository: Repository<Organization>;
-  let organizationUsersRepository: Repository<OrganizationUser>;
+  const userRepository: Repository<User> = app.get('UserRepository');
+  const organizationRepository: Repository<Organization> = app.get('OrganizationRepository');
+  const organizationUsersRepository: Repository<OrganizationUser> = app.get('OrganizationUserRepository');
 
-  userRepository = app.get('UserRepository');
-  organizationRepository = app.get('OrganizationRepository');
-  organizationUsersRepository = app.get('OrganizationUserRepository');
+  // userRepository = app.get('UserRepository');
+  // organizationRepository = app.get('OrganizationRepository');
+  // organizationUsersRepository = app.get('OrganizationUserRepository');
 
-  organization = organization || await organizationRepository.save(organizationRepository.create({ 
-    name: 'test org', 
-    createdAt: new Date(), 
-    updatedAt: new Date() 
-  }));
+  organization =
+    organization ||
+    (await organizationRepository.save(
+      organizationRepository.create({
+        name: 'test org',
+        createdAt: new Date(),
+        updatedAt: new Date(),
+      })
+    ));
 
-  const user = await userRepository.save(userRepository.create({ 
-    firstName: firstName || 'test', 
-    lastName: lastName || 'test', 
-    email: email || 'dev@tooljet.io', 
-    password: 'password', 
-    organization, 
-    createdAt: new Date(), 
-    updatedAt: new Date(), 
-  }));
+  const user = await userRepository.save(
+    userRepository.create({
+      firstName: firstName || 'test',
+      lastName: lastName || 'test',
+      email: email || 'dev@tooljet.io',
+      password: 'password',
+      organization,
+      createdAt: new Date(),
+      updatedAt: new Date(),
+    })
+  );
 
-  const orgUser = await organizationUsersRepository.save(organizationUsersRepository.create({ 
-    user: user, 
-    organization, 
-    role: role || 'admin', 
-    createdAt: new Date(), 
-    updatedAt: new Date()
-  }));
+  const orgUser = await organizationUsersRepository.save(
+    organizationUsersRepository.create({
+      user: user,
+      organization,
+      role: role || 'admin',
+      createdAt: new Date(),
+      updatedAt: new Date(),
+    })
+  );
 
-  return { organization, user, orgUser }
+  return { organization, user, orgUser };
 }
 
 export async function createDataSource(nestInstance, { name, application, kind, options }: any) {
-  let dataSourceRepository: Repository<DataSource>;
-  dataSourceRepository = nestInstance.get('DataSourceRepository');
+  const dataSourceRepository: Repository<DataSource> = nestInstance.get('DataSourceRepository');
+  // dataSourceRepository = nestInstance.get('DataSourceRepository');
 
-  return await dataSourceRepository.save(dataSourceRepository.create({
-    name,
-    options,
-    app: application,
-    kind,
-    createdAt: new Date(), 
-    updatedAt: new Date()
-  }));
+  return await dataSourceRepository.save(
+    dataSourceRepository.create({
+      name,
+      options,
+      app: application,
+      kind,
+      createdAt: new Date(),
+      updatedAt: new Date(),
+    })
+  );
 }
-
 
 export async function createDataQuery(nestInstance, { application, kind, dataSource, options }: any) {
-  let dataQueryRepository: Repository<DataQuery>;
-  dataQueryRepository = nestInstance.get('DataQueryRepository');
+  const dataQueryRepository: Repository<DataQuery> = nestInstance.get('DataQueryRepository');
+  // dataQueryRepository = nestInstance.get('DataQueryRepository');
 
-  return await dataQueryRepository.save(dataQueryRepository.create({
-    options,
-    app: application,
-    kind,
-    dataSource,
-    createdAt: new Date(), 
-    updatedAt: new Date()
-  }));
+  return await dataQueryRepository.save(
+    dataQueryRepository.create({
+      options,
+      app: application,
+      kind,
+      dataSource,
+      createdAt: new Date(),
+      updatedAt: new Date(),
+    })
+  );
 }
-
